+++ src/funcs/operations.rs
@@ -1,4 +1,4 @@
-pub fn add(stack: &mut Vec<f64>, cmd: Vec<&str>, vars: &mut super::super::Vars, b: &str, l: u32) {
+pub fn add(stack: &mut Vec<u8>, cmd: Vec<&str>, vars: &mut super::super::Vars, b: &str, l: u32) {
     if cmd.len() > 1 {
         stack.push(vars.lx + vars.rv);
     } else {
@@ -11,7 +11,7 @@ pub fn add(stack: &mut Vec<f64>, cmd: Vec<&str>, vars: &mut super::super::Vars, 
     }
 }
 
-pub fn sub(stack: &mut Vec<f64>, b: &str, l: u32) {
+pub fn sub(stack: &mut Vec<u8>, b: &str, l: u32) {
     if stack.len() < 2 {
         super::errors::stack_len_error(b, l);
     }
@@ -19,7 +19,7 @@ pub fn sub(stack: &mut Vec<f64>, b: &str, l: u32) {
     stack.push(result);
 }
 
-pub fn mul(stack: &mut Vec<f64>, cmd: Vec<&str>, vars: &mut super::super::Vars, b: &str, l: u32) {
+pub fn mul(stack: &mut Vec<u8>, cmd: Vec<&str>, vars: &mut super::super::Vars, b: &str, l: u32) {
     if cmd.len() > 1 {
         stack.push(vars.lx * vars.rv);
     } else {
@@ -31,7 +31,7 @@ pub fn mul(stack: &mut Vec<f64>, cmd: Vec<&str>, vars: &mut super::super::Vars, 
     }
 }
 
-pub fn div(stack: &mut Vec<f64>, b: &str, l: u32) {
+pub fn div(stack: &mut Vec<u8>, b: &str, l: u32) {
     if stack.len() < 2 {
         super::errors::stack_len_error(b, l);
     }
@@ -39,7 +39,7 @@ pub fn div(stack: &mut Vec<f64>, b: &str, l: u32) {
     stack.push(result)
 }
 
-pub fn sqr(stack: &mut Vec<f64>, cmd: Vec<&str>, vars: &mut super::super::Vars, b: &str, l: u32) {
+pub fn sqr(stack: &mut Vec<u8>, cmd: Vec<&str>, vars: &mut super::super::Vars, b: &str, l: u32) {
     if cmd.len() == 1 {
         if stack.is_empty() {
             super::errors::stack_len_error(b, l);
@@ -60,7 +60,7 @@ pub fn sqr(stack: &mut Vec<f64>, cmd: Vec<&str>, vars: &mut super::super::Vars, 
     }
 }
 
-pub fn sqrt(stack: &mut Vec<f64>, cmd: Vec<&str>, vars: &mut super::super::Vars, b: &str, l: u32) {
+pub fn sqrt(stack: &mut Vec<u8>, cmd: Vec<&str>, vars: &mut super::super::Vars, b: &str, l: u32) {
     if cmd.len() == 1 {
         if stack.is_empty() {
             super::errors::stack_len_error(b, l);
@@ -82,7 +82,7 @@ pub fn sqrt(stack: &mut Vec<f64>, cmd: Vec<&str>, vars: &mut super::super::Vars,
     }
 }
 
-pub fn round(stack: &mut Vec<f64>, cmd: Vec<&str>, vars: &mut super::super::Vars, b: &str, l: u32) {
+pub fn round(stack: &mut Vec<u8>, cmd: Vec<&str>, vars: &mut super::super::Vars, b: &str, l: u32) {
     if cmd.len() > 1 {
         if cmd[1].trim() == "lx" || cmd[1].trim() == "rv" {
             if cmd[1].trim() == "lx" {
@@ -101,7 +101,7 @@ pub fn round(stack: &mut Vec<f64>, cmd: Vec<&str>, vars: &mut super::super::Vars
     }
 }
 
-pub fn avg(stack: &mut Vec<f64>, b: &str, l: u32) {
+pub fn avg(stack: &mut Vec<u8>, b: &str, l: u32) {
     if stack.is_empty() {
         super::errors::stack_len_error(b, l);
     }
@@ -114,7 +114,7 @@ pub fn avg(stack: &mut Vec<f64>, b: &str, l: u32) {
     stack.push(total / i)
 }
 
-pub fn cmp(stack: &mut Vec<f64>, b: &str, l: u32) {
+pub fn cmp(stack: &mut Vec<u8>, b: &str, l: u32) {
     if stack.len() < 2 {
         super::errors::stack_len_error(b, l);
     }
@@ -170,7 +170,7 @@ pub fn split(cmd: Vec<&str>, statement: &str, vars: &mut super::super::Vars, b: 
 // vectors
 
 pub fn vec_ops(
-    stack: &mut Vec<f64>,
+    stack: &mut Vec<u8>,
     cmd: Vec<&str>,
     statement: &str,
     vars: &mut super::super::Vars,
